feat: Add pyproject.toml configuration file
Add pyproject.toml configuration file to the project. This file defines the
build system requirements and project metadata. The build system requires
setuptools version 61.0 or higher, and uses setuptools.build_meta as the
build backend. The project name is "mgdl-simple-cache" with version "0.0.1-rc1".
The description provides an overview of the Simple Cache library, which is a
lightweight cache manager designed to simplify caching operations using Deta
Base. The readme file is "README.md" and the author is Adaías Magdiel with email
address "eu@adaiasmagdiel.com". The license file is "LICENSE" and the project
has various classifiers and keywords to describe its purpose and features.
The required Python version is 3.6 or higher, and the only dependency is the
"deta" library. Additionally, there is an optional development dependency for
running tests, which is "pytest".