feat: Add badges to README
Adding badges to the README to improve visibility and provide important information about the project. This commit adds two badges: one displaying the PyPI version of the package and another displaying the license information. These badges help users quickly understand the status and licensing of the project without needing to search for this information in other places.